Extract makeEnv
