+++ cdk/src/main/java/com/issue/importer/cdk/util/Validations.java
@@ -1,4 +1,4 @@
-package com.issue.importer.cdk;
+package com.issue.importer.cdk.util;
 
 import software.amazon.awscdk.App;
